Extract monthly equivalent calculations to mixin
To use in new collator class
